FIX BUG in Weather Info Section
When the weather API was accessed after 21:00 the app was not able to present the fifth day, because the always went out of boundry of the last record that API provided.

Issue was fixed, by making sure that the highest possible index that could be accessed was 39 as the whole JSON response returns JSON file with 40 items. Thus, index 40 was trying to reach non existent record which in turn casued the error.